requirements.txt
google-genai
biopython
python-dotenv
